% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesbench_cfg.R
\name{bayesbench_cfg}
\alias{bayesbench_cfg}
\alias{read_bayesbench_cfg_from_file}
\alias{write_bayesbench_cfg_to_file}
\title{Bayesbench config}
\usage{
bayesbench_cfg(...)

read_bayesbench_cfg_from_file(file_path)

write_bayesbench_cfg_to_file(cfg, file_path)
}
\arguments{
\item{...}{Bayesbench parameters to set.}

\item{file_path}{File path to bayesbench configuration file.}
}
\description{
Bayesbench config
}
\details{
Creates a Bayesbench configuration file. The only
}
\examples{
file_path <- system.file("extdata", "examples", "test_8schools_advi.yml", package = "bayesbenchr")
x <- read_bayesbench_cfg_from_file(file_path)

}
