% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesbench_output.R
\name{bayesbench_output}
\alias{bayesbench_output}
\title{Create a Bayesbench ouput object}
\usage{
bayesbench_output(cfg, posterior, posterior_predictive = NULL,
  diagnostics = NULL, inference_engine_content = NULL,
  output_log = NULL)
}
\arguments{
\item{cfg}{The config object to return}

\item{posterior}{A list of posterior, one array per parameter}

\item{posterior_predictive}{A list of posterior predictives, one array per parameter predictive}

\item{inference_engine_content}{a list of inference_engine specific content (these need stored be handled separately)}

\item{log_p}{The log target density (log posterior mostly) for each draw}

\item{log_q}{The log proposal density (log approximate posterior) for each draw}

\item{text_log}{eventual output during inference engine run}
}
\description{
Create a Bayesbench ouput object
}
