% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bayesbench.R
\name{bayesbench_run}
\alias{bayesbench_run}
\alias{bayesbench_bash_run}
\title{Run Bayesbench}
\usage{
bayesbench_run(cfg)

bayesbench_bash_run(cfg, bash_folder = "temp_bayesbench",
  processes = 1)
}
\arguments{
\item{cfg}{a \code{bayesbench_cfg} object or a list of \code{bayesbench_cfg} object.}

\item{bash_folder}{Folder to write bash scripts to}

\item{cores}{how many files to run (to start as differet processes)}
}
\description{
Run Bayesbench for a given set of configurations
}
\examples{
path <- system.file("extdata", "examples", "test_8schools_advi.yml", package = "bayesbenchr")
cfg <- read_bayesbench_cfg_from_file(path)
# results <- bayesbench_run(cfg)

}
