% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesbench_cfg.R
\name{expand_bayesbench_cfg_to_job_cfgs}
\alias{expand_bayesbench_cfg_to_job_cfgs}
\title{Expand \code{bayesbench_cfg} objects to \code{bayesbench_job_cfg}s}
\usage{
expand_bayesbench_cfg_to_job_cfgs(cfg)
}
\arguments{
\item{cfg}{a list of \code{bayesbench_cfg} to expand.}
}
\description{
Expand \code{bayesbench_cfg} objects to \code{bayesbench_job_cfg}s
}
\details{
job configs are individual configurations that can be run independently
}
