% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesbench_output.R
\name{read_bayesbench_outputs}
\alias{read_bayesbench_outputs}
\title{Read Bayesian outputs}
\usage{
read_bayesbench_outputs(x)
}
\arguments{
\item{x}{either file paths or a cfg object}
}
\description{
Read Bayesian outputs
}
\details{
Read in bayesbench output jobs
}
