% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bayesbench.R
\name{run_bayesbench}
\alias{run_bayesbench}
\title{Run Bayesbench}
\usage{
run_bayesbench(cfg)
}
\arguments{
\item{cfg}{a \code{bayesbench_cfg} object or a list of \code{bayesbench_cfg} object.
path <- system.file("extdata", "examples", "test_8schools_advi.yml", package = "bayesbenchr")
cfg <- read_bayesbench_cfg_from_file(path)}
}
\description{
Run Bayesbench
}
