% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie_stan_advi.R
\name{stan_advi}
\alias{stan_advi}
\title{Run Stan VB functionality}
\usage{
stan_advi(cfg)
}
\arguments{
\item{cfg}{a \code{bayesbench_cfg} config object}
}
\value{
a \code{bayesbench_output} object
}
\description{
Run Stan VB functionality
}
\details{
Runs Stans VB inferences.
}
