% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie_stan.R
\name{stan_optimizing}
\alias{stan_optimizing}
\title{Run Stan optimizing/laplace}
\usage{
stan_optimizing(cfg)
}
\arguments{
\item{cfg}{a \code{bayesbench_cfg} config object}
}
\value{
a \code{bayesbench_output} object
}
\description{
Run Stan optimizing/laplace
}
