% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie_stan_advi.R
\name{stan_sampling}
\alias{stan_sampling}
\title{Run Stan MCMC}
\usage{
stan_sampling(cfg)
}
\arguments{
\item{cfg}{a \code{bayesbench_cfg} config object}
}
\value{
a \code{bayesbench_output} object
}
\description{
Run Stan MCMC
}
