% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie_stan_advi.R
\name{stan_vb}
\alias{stan_vb}
\title{Run Stan VB}
\usage{
stan_vb(cfg)
}
\arguments{
\item{cfg}{a \code{bayesbench_cfg} config object}
}
\value{
a \code{bayesbench_output} object
}
\description{
Run Stan VB
}
\details{
Runs Stans VB inferences.
}
