% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesbench_output.R
\name{write_bayesbench_outputs}
\alias{write_bayesbench_outputs}
\title{Write bayesbench outputs to file}
\usage{
write_bayesbench_outputs(bayesbench_outputs, output_directory = NULL,
  output_type = NULL)
}
\description{
Write bayesbench outputs to file
}
\details{
Will write \code{bayesbench_outputs} to file(s) according to the \code{bayesbench_outputs} configs.
Specifying \code{output_directory} or \code{output_type} will update the config with these values
and then write it to file(s).
}
