% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesbench_output.R
\name{bayesbench_output}
\alias{bayesbench_output}
\title{Create a Bayesbench ouput object}
\usage{
bayesbench_output(cfg, posterior, posterior_predictive = NULL,
  log_p = NULL, log_g = NULL, method_diagnostic = NULL,
  text_log = NULL)
}
\arguments{
\item{cfg}{The config object to return}

\item{posterior}{A list of posterior, one array per parameter}

\item{posterior_predictive}{A list of posterior predictives, one array per parameter predictive}

\item{log_p}{The log target density (log posterior mostly) for each draw}

\item{method_diagnostic}{a list of method specific diagnostics (these need sto be handled separately)}

\item{text_log}{eventual output during inference engine run}

\item{log_q}{The log proposal density (log approximate posterior) for each draw}
}
\description{
Create a Bayesbench ouput object
}
