% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesbench_output.R
\name{bayesbench_output}
\alias{bayesbench_output}
\title{Create a Bayesbench ouput object}
\usage{
bayesbench_output(cfg, posterior, posterior_predictive = NULL,
  log_p = NULL, log_g = NULL, method_diagnostic = NULL,
  text_log = NULL)
}
\description{
Create a Bayesbench ouput object
}
